Fix the overflow issue caused by the mem index type being int32 in the decode att operator. (#1013)

+++ lightllm/models/llama/triton_kernel/gqa_flash_decoding_stage1.py
@@ -76,7 +76,7 @@ def _fwd_kernel_flash_decode_stage1(
             Req_to_tokens + stride_req_to_tokens_b * cur_batch_req_idx + offs_n_new,
             mask=offs_n_new < cur_batch_end_index,
             other=0,
-        )
+        ).to(tl.int64)
         off_k = k_loc[None, :] * stride_kbs + cur_kv_head * stride_kh + offs_d[:, None]
         k = tl.load(K + off_k, mask=offs_n_new[None, :] < cur_batch_end_index, other=0.0)
         att_value = tl.dot(q, k)
